Fixed lint violations in the migrated analytics code
ref https://linear.app/ghost/issue/PLA-122/co-locate-stats-into-appsadmin-as-the-analytics-domain

- the stats package linted under a looser config, so under Admin's type-aware rules the same code reports unsafe member access, unsafe arguments, base-to-string and floating promises
- narrowed types at the boundaries that produce them (JSON.parse results, Recharts payloads, untyped API values) instead of disabling rules, so analytics holds the same strictness as the rest of Admin
- moved a hook and a constant out of component files to satisfy react-refresh's single-export rule
