+++ apps/admin/src/analytics/providers/analytics-provider.tsx
@@ -1,48 +1,10 @@
-import {type Config, useBrowseConfig} from '@tryghost/admin-x-framework/api/config';
-import {type ReactNode, createContext, useContext, useState} from 'react';
+import {type ReactNode, useState} from 'react';
+import {AnalyticsContext} from '@/analytics/providers/analytics-context';
 import {STATS_DEFAULT_RANGE_KEY, STATS_RANGE_OPTIONS} from '@/analytics/utils/constants';
-import {type Setting, useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
-import {type StatsConfig, useTinybirdToken} from '@tryghost/admin-x-framework';
+import {useBrowseConfig} from '@tryghost/admin-x-framework/api/config';
+import {useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
 import {useBrowseSite} from '@tryghost/admin-x-framework/api/site';
-
-interface AnalyticsData {
-    // --- View-state: owned by this provider (the slim end-state) ---
-    range: number;
-    setRange: (value: number) => void;
-    selectedNewsletterId: string | null;
-    setSelectedNewsletterId: (id: string | null) => void;
-
-    // --- TODO(PLA-192): framework-data below is a temporary passthrough. ---
-    // config/site/settings/tinybirdToken all duplicate data the shell already
-    // exposes via admin-x-framework hooks (FrameworkProvider + AppProvider wrap
-    // the whole admin tree). Hoisting these out of the analytics context lets
-    // this provider shrink to just the view-state above. Deferred to avoid
-    // churning the ~18 consumer call sites in this spike.
-    data: Config | undefined;
-    site: {
-        url?: string;
-        icon?: string;
-        title?: string;
-    };
-    statsConfig: StatsConfig | undefined;
-    tinybirdToken: string | undefined;
-    isLoading: boolean;
-    settings: Setting[];
-}
-
-const AnalyticsContext = createContext<AnalyticsData | undefined>(undefined);
-
-export const useAnalytics = () => {
-    const context = useContext(AnalyticsContext);
-    if (context === undefined) {
-        throw new Error('useAnalytics must be used within an AnalyticsProvider');
-    }
-    return context;
-};
-
-// TODO(PLA-124): Back-compat alias. Migrate the ~18 `useGlobalData` call sites in
-// src/analytics to `useAnalytics`, then remove this export.
-export const useGlobalData = useAnalytics;
+import {useTinybirdToken} from '@tryghost/admin-x-framework';
 
 const AnalyticsProvider = ({children}: { children: ReactNode }) => {
     // --- View-state (this provider's real responsibility) ---
@@ -72,7 +34,7 @@ const AnalyticsProvider = ({children}: { children: ReactNode }) => {
     const isLoading = isGhostLoading || isTinybirdLoading;
 
     if (error) {
-        throw error;
+        throw error instanceof Error ? error : new Error('Failed to load analytics framework data');
     }
 
     // Extract site data
