+++ apps/admin/src/analytics/views/Stats/Growth/components/growth-sources.tsx
@@ -7,7 +7,7 @@ import {LucideIcon, formatNumber} from '@tryghost/shade/utils';
 import {centsToDollars} from '@tryghost/shade/app';
 import {getFaviconDomain, getSymbol, useAppContext} from '@tryghost/admin-x-framework';
 import {getPeriodText} from '@/analytics/utils/chart-helpers';
-import {useGlobalData} from '@/analytics/providers/analytics-provider';
+import {useGlobalData} from '@/analytics/providers/analytics-context';
 import {useMrrHistory} from '@tryghost/admin-x-framework/api/stats';
 import {useTopSourcesGrowth} from '@/analytics/hooks/use-top-sources-growth';
