+++ apps/admin/src/analytics/views/Stats/Growth/growth.tsx
@@ -16,7 +16,7 @@ import {centsToDollars} from '@tryghost/shade/app';
 import {getClickHandler} from '@/analytics/utils/url-helpers';
 import {getPeriodText} from '@/analytics/utils/chart-helpers';
 import {useAppContext} from '@tryghost/admin-x-framework';
-import {useGlobalData} from '@/analytics/providers/analytics-provider';
+import {useGlobalData} from '@/analytics/providers/analytics-context';
 import {useGrowthStats} from '@/analytics/hooks/use-growth-stats';
 import {useNavigate, useSearchParams} from '@tryghost/admin-x-framework';
 import {useTopPostsStatsWithRange} from '@/analytics/hooks/use-top-posts-stats-with-range';
