+++ apps/admin/src/analytics/views/Stats/Overview/components/latest-post.tsx
@@ -6,7 +6,7 @@ import {PostShareModal} from '@tryghost/shade/posts-stats';
 import {type Post, getPostMetricsToDisplay} from '@tryghost/admin-x-framework';
 import {getPostDestination} from '@/analytics/utils/url-helpers';
 import {useAppContext, useNavigate} from '@tryghost/admin-x-framework';
-import {useGlobalData} from '@/analytics/providers/analytics-provider';
+import {useGlobalData} from '@/analytics/providers/analytics-context';
 
 // Import the interface from the hook
 import {type LatestPostWithStats} from '@/analytics/hooks/use-latest-post-stats';
