+++ apps/admin/src/analytics/views/Stats/components/date-range-select.tsx
@@ -2,7 +2,7 @@ import React, {useEffect} from 'react';
 import {LucideIcon} from '@tryghost/shade/utils';
 import {STATS_RANGES, STATS_RANGE_OPTIONS} from '@/analytics/utils/constants';
 import {Select, SelectContent, SelectGroup, SelectItem, SelectLabel, SelectTrigger, SelectValue} from '@tryghost/shade/components';
-import {useGlobalData} from '@/analytics/providers/analytics-provider';
+import {useGlobalData} from '@/analytics/providers/analytics-context';
 
 interface DateRangeSelectProps {
     excludeRanges?: (keyof typeof STATS_RANGES)[];
